HIG-778: Restrict guest operations from the backend + fix security vuln (#1070)
Guests should be able to view a session but not be able to modify it - this backend change ensures guests cannot perform actions they are not allowed to do. Also did the minimum on the frontend to keep the guest experience functional, but more changes are coming.

Also a little bit of cleanup, including fixing a security vulnerability where anyone can post a comment on any session masquerading as someone else.

Testing: Load session correctly as both admin and guest. Comment and bookmark successfully as admin. Guest attempting to bookmark a session fails.

Future changes:

    Hide operations on the frontend that a guest cannot do
    Allow guests to comment using a dummy guest account in the backend
    Sharable errors
    Use random string instead of xid
    Secure session URL's