HIG-995: Pull field filters into a separate function (#1192)
Trying to split the DB refactor up into pieces that are easy to understand. First step is splitting the big query into smaller, easier to understand units, before refactoring each unit.

There is some logic in the endpoint to issue a query to fetch all of the field ID's satisfying a condition the user specified, and then include/exclude them from the final DB query using a WHERE clause. Let's split all of this logic into a separate function. Note I'm not trying to make any functional changes in this PR.