HIG-995: Improve perf and readability of filtered sessions query (#1244)
Logic before:
A1. Fetch all fields across all orgs that match the inputs the user selected, get the field ID's (potentially thousands of ID's, for example with Impira's query for emails that have '@gmail.com' or '@hotmail.com').
A2. Get the top 20 sessions that have those ID's.

Logic now:
B1. Get the top 20 sessions that have fields that match the inputs the user selected.

According to explain, B1 is 20x faster than A2 to begin with, and on top of that we don't need to run A1 at all, so this should be significantly faster.

Note SQL injection is possible with this code - there are many places in the codebase where this is possible, so I have chosen to defer this problem and fix all of them together later.