Generate secure session ID from firstload (#2265)
A collision is pretty much impossible (https://neilmadden.blog/2018/08/30/moving-away-from-uuids/), but if it does happen, the backend will fail to create the session as there is a unique index on sessionSecureID.

This does mean that a malicious person can modify the JavaScript in firstload to set the secure ID for their session to whatever they want. I am unable to think of a problem with this. Note that even before this change, people can add payloads to other people's sessions if they know the (non-secure) ID for that session.