[HIG-2113] show active network requests (#2408)
show which network and console events are currently
active based on the player time. shows the event times in the resource page as well.
when clicking the player timeline, ensure the network events  jump
to the appropriate timeframe. choosing a network resource also now
moves the player to the appropriate time of the request.