[HIG-2321] stop storing uncompressed session data (#2642)
Only store compressed data and use the compressed data
as a source for other parts of the code that read the uncompressed data.

In the backend, some processing code needs to read the session data, so add
a step after retrieving objects from S3 to decompress.

When the frontend loads a session, we ask the backend for direct URLs
to the compressed payload only. However, the backend only retrieves these
when the browser natively supports brotli. When it does not, the browser calls
`GetSessionPayload` to retrieve decompressed resources.

This will allow us to delete old uncompressed objects from s3 to save storage costs.

Testing: chrome canary with brotli disabled and object storage enabled,
calls `GetSessionPayload` query which decompressed the compressed resources
correctly in the backend.