add s3 uncompressed object archival script (#2653)
take uncompressed objects and migrate them to glacier deep archive storage tier for sessions that have
compressed objects for all file types.

the script is multiprocessed to allow one process to find sessions with uncompressed and compressed files (to archive the uncompressed ones) and feeds these to a pool of subprocesses workers.

we would run this on highlight sessions only to start. the glacier `DEEP_ARCHIVE` storage class is 5% of the standard storage price, and we can later delete the archived uncompressed data. glacier archived files have to be unarchived to make them accessible again and takes up to 12 hours.