Fix client reset logic (#2973)
* Fix fetch reset

We don't want to reset window.fetch because that overrides the changes we make in initializeFetchListener. We want to reset window._fetchProxy otherwise our logic is never hit again.

* Drop initMembers call

We only need to call `initMembers` when initializing the client. It has problematic side effects of calling it on reset. e.g. calling `this.listeners = []` isn't correct because we never teardown those listeners.

* Bump client version

* Remove conditional for reloaded sessions

* Bump firstload version