fix upgrade-plan page not loading auth correctly (#3064)
The `upgrade-plan` route would require a refresh to not load the authorization fallback card.
This occurred because the original load of the frontend on `/` would not have a project/workspace, and would
not be able to load a workspace role for the user. Changes to the URL would not trigger the router updates
due to a bug in react-router 5 with react 18 - https://linear.app/highlight/issue/HIG-2647

Until then, this provides a workaround using window pathname.