update kafka worker concurrency (#3122)
## Summary

We're seeing an ok overall backlog but a very high max partition offset (most messages are in 5 of 2048 partitions). Want to try upping worker concurrency and prefetch amount to see if that helps.

## How did you test this change?

Local backend deploy.

## Are there any deployment considerations?

Will monitor ECS container memory usage to ensure we do not exceed memory limits.