feat(HIG-2887): update the search result list UI (#3241)
## Summary

- Updates the general look of the left nav for error pages, introducing
a Badge component (see
[Figma](https://www.figma.com/file/pAHlb8IFmVSewPJcphbjIH/Component-Library?node-id=7%3A653&viewport=-28%2C1487%2C1.13))
- Adds a localStorage `highlight-enable-staff-view` param to turn off
highlight staff view (the histogram overview in the left nav both for
errors and sessions was not visible to anyone but highlight admins).

## How did you test this change?

storybook + [render](https://frontend-pr-3241.onrender.com/)

## Are there any deployment considerations?

1. At the moment counts for affected users and sessions are stubs. Will
be fixed after adding necessary backend changes.
2. This PR changes the height of the histogram in the left nav for
sessions as well. This change does not affect our customers, because the
histogram was gated for highlight staff in the first place.