hotfix: remove useEffect watching the panels (#3301)
## Summary

Context: Vera from enterneverland.com cannot open the right panel on a
pretty large screen (1247 x 1810).


https://app.intercom.com/a/inbox/gm6369ty/inbox/shared/all/conversation/41135?view=List

I cannot reproduce this behavior, but believe this is due to a useEffect
that should be closing the right panel if they are both open on a small
enough screen. Moving the logic that does that to event handlers on
buttons.

Session on highlight showcasing the behavior:


https://app.highlight.run/1/sessions/LO7wtMT9JSxMmJmAS3XmDLlOISDw?page=1&query=and%7C%7Ccustom_processed%2Cis%2Ctrue%7C%7Ccustom_created_at%2Cbetween_date%2C30%20days%7C%7Cuser_identifier%2Cis%2Cvera%40enterneverland.com

## How did you test this change?

Chrome dev tools. 

## Are there any deployment considerations?

No