migrate home dashboards to new influxdb charts (#3292)
## Summary

Brings back #3286 with fixes.

Now that we have 90+ days of influxdb data for session and error counts, use the influxdb-based charts
for all new projects and migrate existing home dashboards to the new chart types.
Provides a way to migrate the home dashboard in the future.

## How did you test this change?

Local loading of the home page for a project on the old dashboard updated it to the new look.

<img width="853" alt="Screenshot 2022-11-04 at 3 35 40 PM" src="https://user-images.githubusercontent.com/1351531/200084643-6cc68f79-b03a-4c66-b1c9-dde8f5a3f3d0.png">


## Are there any deployment considerations?

No. Migration will only affect the home dashboard definitions as users visit their project home page.