fix cache options for client bundle (#3397)
## Summary
- Set `Cache-Control` and `Content-Type` headers for uploaded JS bundle
<!--
Ideally, there is an attached Linear ticket that will describe the
"why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

## How did you test this change?
- updated Cache-Control and Content-Type headers for current bundle and
verified
  - compression issue was no longer flagged in Lighthouse report
- cache settings were still flagged but showed the 1 hour TTL - my
understanding from the Lighthouse docs is that the page is refreshed
multiple times and anything that should be cached and isn't is flagged
here (e.g. because we have the `?123` in our URL). If this is still an
issue, may have to increase `max-age` or mark immutable

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?
- no
<!--
 Backend - Do we need to consider migrations or backfilling data?
-->