add height integration (#3472)
## Summary

<!--
Ideally, there is an attached Linear ticket that will describe the
"why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

This adds a [Height](https://height.app/) integration that works pretty
similar to Clickup (see #3246).

Most notably is the introduction of `IntegrationWorkspaceMapping` which
stores the oauth information. This is similar to the polymorphic table
that #3246 introduced called `IntegrationProjectMapping` which is
project -> external id (where external id is a clickup task or a height
task). In theory, if we add a new integration, we shouldn't need any
more database changes

## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

* Verified we can integrate Highlight and Height via OAuth
* Verified we can map projects to Height workspaces (currently only one
allowed)
* Verified we can create new Height tasks from sessions and errors and
we can attach a Height list to the task

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

* Ensure `HEIGHT_CLIENT_ID` and `HEIGHT_CLIENT_SECRET` are set in
doppler
* Contact Height to allow other customers to use this integration
(currently only [test workspace](https://height.app/52uHzTBUU3) is
authorized). See

> Workspace - Workspace that will be linked to your integration. Your
integration will initially be only only available to that workspace.
Contact us when it’s ready to be released to Height’s customers

* Ideally, we should consolidate some of the clickup code to use some of
the generic code introduced in this PR