improve kafka log batching and highlight-py sdk (#4185)
## Summary

* Allow highlight python sdk to override OTLP endpoint.
* Batch messages from multiple partitions into a shared in-memory
buffer.
* Don't write log messages as an attribute.
     
currently, each public-worker container runs 8 goroutines for the
batched topic. each goroutine gets a set of subset of partitions
assigned to it. as messages are consumed by the goroutines, we should
aggregate the LogRows together across these partitions so that the
batching can be more substantial. without this change, each goroutine
had its own in-memory buffer which would not get many messages (and
actually could lead to dropping messages because a goroutine would flush
its partition's messages as the front of the topic, while in reality
there may be other unflushed messages)

## How did you test this change?

Local deploy with debugger.

## Are there any deployment considerations?

New highlight python sdk version.