allow changing placeholder in ErrorBoundary dialog (#4354)
## Summary

The ReportDialog component has all strings parametrized, which es really
helpful specially for those requiring the component to be shown to users
in a different language. The placeholder in the comments is the only one
that was still hardcoded:


![image](https://user-images.githubusercontent.com/4401263/220734833-8e3515a9-89f3-4fc2-95d8-c451e36de96f.png)

## How did you test this change?

I haven't been able to test the change yet.

## Are there any deployment considerations?

Since this adds a new optional argument only to ReportDialogOptions,
there shouldn't be any backwards compatibility issues. I imagine there
might be some scenarios where the placeholder now shows empty for some
users that added customization, but not sure.