fix recording in iframes without highlight in the parent window (#4508)
## Summary

Highlight does not record correctly in an iframe because of the
cross-origin setting. Before this change,
when Highlight is added to an iframe with the intention to record just
the iframe, we would assume that
the iframe was emitting events to the parent while there was no
Highlight instance in the parent.

Since previously we would pass `recordCrossOriginIframes: true` to
rrweb, the following logic in rrweb would never emit events in an
iframe.

```ts
  const inEmittingFrame = recordCrossOriginIframes
    ? window.parent === window
    : true;
```

## How did you test this change?

Buttons page iframe recording.
![Screenshot 2023-03-06 at 12 55 09
PM](https://user-images.githubusercontent.com/1351531/223228559-53f22177-a6c0-47e8-aa34-d59505bb8fef.png)


## Are there any deployment considerations?

New highlight.run minor version bump.